checkout progress: report as quint64 type

+++ libGitWrap/Events/IGitEvents.hpp
@@ -59,8 +59,8 @@ namespace Git
                                      const DiffFile& target,
                                      const DiffFile& workdir ) = 0;
         virtual void checkoutProgress( const QString& path,
-                                       quint32 total,
-                                       quint32 completed ) = 0;
+                                       quint64 total,
+                                       quint64 completed ) = 0;
         virtual void doneCheckout() = 0;
     };
 }
